@@ -1,22 +1,23 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.4.30'
+    ext.kotlin_version = '1.5.10'
     repositories {
+        maven { url = 'https://nexus3.mystery0.vip/repository/maven-public/' }
         google()
-        jcenter()
+        mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.1.2'
+        classpath 'com.android.tools.build:gradle:7.0.0-beta04'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-		classpath 'vip.mystery0.tools:autoversion:1.0.7'
     }
 }
 
 allprojects {
     repositories {
+        maven { url = 'https://nexus3.mystery0.vip/repository/maven-public/' }
         google()
-        jcenter()
+        mavenCentral()
     }
 }
 
