@@ -2,6 +2,7 @@ package pw.janyo.whatanime.ui.activity
 
 import android.content.Intent
 import androidx.databinding.ViewDataBinding
+import com.orhanobut.logger.Logger
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import pw.janyo.whatanime.R
 import pw.janyo.whatanime.base.WABaseActivity
@@ -11,7 +12,6 @@ import pw.janyo.whatanime.config.inBlackList
 import pw.janyo.whatanime.config.useServerCompress
 import pw.janyo.whatanime.model.response.StatisticsResponse
 import pw.janyo.whatanime.viewModel.TestViewModel
-import vip.mystery0.logs.Logs
 import vip.mystery0.rx.PackageDataObserver
 import vip.mystery0.tools.ResourceException
 
@@ -38,7 +38,7 @@ class ReceiveShareActivity : WABaseActivity<ViewDataBinding>(null) {
             override fun error(data: StatisticsResponse?, e: Throwable?) {
                 super.error(data, e)
                 if (e !is ResourceException) {
-                    Logs.wtf("error: ", e)
+                    Logger.wtf("error: ", e)
                 }
                 doNext()
             }
