@@ -1,5 +1,7 @@
 package pw.janyo.whatanime.ui.adapter
 
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.inject
 import pw.janyo.whatanime.R
 import pw.janyo.whatanime.databinding.ItemHistoryBinding
 import pw.janyo.whatanime.handler.HistoryItemListener
@@ -13,23 +15,25 @@ import vip.mystery0.tools.utils.toDateTimeString
 import java.io.File
 import java.text.DecimalFormat
 
-class HistoryRecyclerAdapter(
-		private val listener: HistoryItemListener
-) : BaseBindingRecyclerViewAdapter<AnimationHistory, ItemHistoryBinding>(R.layout.item_history) {
+class HistoryRecyclerAdapter :
+    BaseBindingRecyclerViewAdapter<AnimationHistory, ItemHistoryBinding>(R.layout.item_history),
+    KoinComponent {
+    private val listener: HistoryItemListener by inject()
 
-	override fun setItemView(binding: ItemHistoryBinding, position: Int, data: AnimationHistory) {
-		val animation = data.result.fromJson<Animation>()
-		binding.handler = listener
-		binding.animation = animation
-		binding.history = data
-		binding.imageView.loadWithoutCache(File(data.cachePath))
-		if (animation.docs.isNotEmpty()) {
-			binding.animationDocs = animation.docs[0]
-			binding.textViewTime.text = data.time.getCalendarFromLong().toDateTimeString()
-			val similarity = "${DecimalFormat("#.0000").format(animation.docs[0].similarity * 100)}%"
-			binding.textViewSimilarity.text = similarity
-		} else {
-			binding.textViewSimilarity.text = "0%"
-		}
-	}
+    override fun setItemView(binding: ItemHistoryBinding, position: Int, data: AnimationHistory) {
+        val animation = data.result.fromJson<Animation>()
+        binding.handler = listener
+        binding.animation = animation
+        binding.history = data
+        binding.imageView.loadWithoutCache(File(data.cachePath))
+        if (animation.docs.isNotEmpty()) {
+            binding.animationDocs = animation.docs[0]
+            binding.textViewTime.text = data.time.getCalendarFromLong().toDateTimeString()
+            val similarity =
+                "${DecimalFormat("#.0000").format(animation.docs[0].similarity * 100)}%"
+            binding.textViewSimilarity.text = similarity
+        } else {
+            binding.textViewSimilarity.text = "0%"
+        }
+    }
 }
